requirements.txt
PyQt6>=6.6.0
PyQt6-WebEngine>=6.6.0
tweepy>=4.14.0
atproto>=0.0.50
Pillow>=10.2.0
keyring>=25.0.0
requests>=2.31.0
packaging>=24.0
python-dotenv>=1.0.0
